feat: transfer filter logic on discover use case
