+++ core/domain/src/main/kotlin/com/divinelink/core/domain/DiscoverMediaUseCase.kt
@@ -3,6 +3,7 @@ package com.divinelink.core.domain
 import com.divinelink.core.commons.domain.DispatcherProvider
 import com.divinelink.core.commons.domain.FlowUseCase
 import com.divinelink.core.data.media.repository.MediaRepository
+import com.divinelink.core.model.discover.DiscoverFilter
 import com.divinelink.core.model.discover.DiscoverParameters
 import com.divinelink.core.model.media.MediaType
 import com.divinelink.core.model.user.data.UserDataResponse
@@ -15,10 +16,22 @@ class DiscoverMediaUseCase(
 ) : FlowUseCase<DiscoverParameters, UserDataResponse>(dispatcher.default) {
 
   override fun execute(parameters: DiscoverParameters): Flow<Result<UserDataResponse>> = flow {
+    val discoverFilters = with(parameters.filters) {
+      buildList {
+        if (genres.isNotEmpty()) {
+          add(DiscoverFilter.Genres(genres.map { it.id }))
+        }
+        language?.let { filter -> add(DiscoverFilter.Language(filter.code)) }
+        country?.let { add(DiscoverFilter.Country(it.code)) }
+        voteAverage?.let { add(voteAverage) }
+        add(DiscoverFilter.MinimumVotes(votes ?: 10))
+      }
+    }.mapNotNull { it }
+
     if (parameters.mediaType == MediaType.TV) {
       repository.discoverTvShows(
         page = parameters.page,
-        filters = parameters.filters,
+        filters = discoverFilters,
       ).collect { result ->
         result.fold(
           onSuccess = { emit(Result.success(it)) },
@@ -28,7 +41,7 @@ class DiscoverMediaUseCase(
     } else {
       repository.discoverMovies(
         page = parameters.page,
-        filters = parameters.filters,
+        filters = discoverFilters,
       ).collect { result ->
         result.fold(
           onSuccess = { emit(Result.success(it)) },
