+++ core/model/src/main/kotlin/com/divinelink/core/model/discover/DiscoverParameters.kt
@@ -5,5 +5,5 @@ import com.divinelink.core.model.media.MediaType
 data class DiscoverParameters(
   val page: Int,
   val mediaType: MediaType,
-  val filters: List<DiscoverFilter>,
+  val filters: MediaTypeFilters,
 )
