+++ feature/discover/src/main/kotlin/com/divinelink/feature/discover/DiscoverUiState.kt
@@ -1,37 +1,9 @@
 package com.divinelink.feature.discover
 
-import com.divinelink.core.model.Genre
-import com.divinelink.core.model.discover.DiscoverFilter
-import com.divinelink.core.model.locale.Country
-import com.divinelink.core.model.locale.Language
+import com.divinelink.core.model.discover.MediaTypeFilters
 import com.divinelink.core.model.media.MediaItem
 import com.divinelink.core.model.media.MediaType
 import com.divinelink.core.model.tab.MediaTab
-
-data class MediaTypeFilters(
-  val genres: List<Genre>,
-  val language: Language?,
-  val country: Country?,
-  val voteAverage: DiscoverFilter.VoteAverage?,
-  val votes: Int?,
-) {
-  companion object {
-    val initial = MediaTypeFilters(
-      genres = emptyList(),
-      language = null,
-      country = null,
-      voteAverage = null,
-      votes = null,
-    )
-  }
-
-  val hasSelectedFilters
-    get() = genres.isNotEmpty() ||
-      language != null ||
-      country != null ||
-      voteAverage != null ||
-      votes != null
-}
 
 data class DiscoverUiState(
   val selectedTabIndex: Int,
